Final code updates and bug fixes
Fixed the code to handle incremental loads correctly.
